+++ src/Hacks/Level/PracticeComplete.cpp
@@ -1,0 +1,24 @@
+/*
+#include "../../Client/Module.hpp"
+#include <Geode/modify/PlayLayer.hpp>
+
+using namespace geode::prelude;
+
+class PracticeComplete : public Module
+{
+    public:
+        MODULE_SETUP(PracticeComplete)
+        {
+            setName("1 Attempt");
+            setID("");
+            setCategory("Level");
+            setDescription("");
+        }
+};
+
+SUBMIT_HACK(PracticeComplete);
+
+class $modify (PlayLayer)
+{
+    
+};*/